configs: nitrogen*: bump kernel and u-boot version for m4 support
Also adding the '-m4' device tree blobs which allow to disable the
peripherals that are used by the Cortex-M4.

The 6x_bootscript is already setup to load the '-m4' dtb if m4enabled
u-boot variable is set to 1.

The FreeRTOS BSP for both Nitrogen6SX and Nitrogen7 can be found here:
https://github.com/boundarydevices/freertos-boundary

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>