package/sg3_utils: fix uclinux build
sg3_utils grep for '^linux' as the host OS to decide if we are
targetting Linux.

However, on noMMU targets, the host OS can be 'uclinux' instead of just
plain 'linux', which make sg3_util confused and thus it forgets to
include some source files in the build, resulting in link failures.

Fix that by ammending the regexp that recognises Linux. We now need to
run autoreconf.

Fixes:
    http://autobuild.buildroot.org/results/79a/79a572f67bafa6a007256a5f22529c2daabd5733/
    http://autobuild.buildroot.org/results/4f7/4f7ea4bb4c4c12ceb9cd08bfb6720386f5bdea0b/

[Peter: fix typos as pointed out by Arnout]
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>