qt: build using old C++ standard
'std::auto_ptr' is deprecated with gcc-6 but is used in qt4 code.
So we have to build using old C++ standard to prevent build failures.

qt4 is not being developed no more, so unfortunately we can't update
obsolete C++ code and have to add this workaround.

Fixes:
http://autobuild.buildroot.net/results/54172d55f39b2fcfa5b7fb4d4ee01566678babbc/
and also Qt build for ARC.

Signed-off-by: Vlad Zakharov <vzakhar@synopsys.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>