package/libgtk3: needs OpenGL-EGL/wayland
Currently, libgtk3 depends on either a full OpenGL stack, or on a
OpenGL-EGL stack provided by mesa3d.

This is because, when using the wayland backend, libgtk3 needs an
OpenGL-EGL stack tht provides the wayland extensions. So far, only
mesa3d would provide those extensions.

But now, those extensions are trickling down to other implementations,
like the Nvidia driver or other (even non-public) implementations.

Change the condition to be on the recently introduced libegl-wayland
option.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>