libiio: do not install systemd service file, no longer available
Since the bump to libiio 0.7, the systemd service file that used to be
in debian/iiod.service no longer exists. The entire debian/ directory
has been removed from the upstream project:

  https://github.com/analogdevicesinc/libiio/commit/5d49f58982b0c15545bc70724304fbc580e6523c

Due to this, the installation of this service file fails, and causes
build failures. To address this, we simply remove the installation of
the systemd service file.

Fixes:

  http://autobuild.buildroot.net/results/ce2fc81466abd9619aa104c96234d1455de3480d/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>