package/qemu: fix build of host variant
The Qemu buildsystem expects that the cpp it is passed can be called
like it were cc: it passes it the '-c' option, which is not allowed by
cpp, but is allowed for cc.

Fix that by overriding CPP when calling configure.

Note: the target variant of Qemu does not build the affected parts (the
PC-BIOS images). This is another problem that should be fixed separately.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>