ts4900-fpga: add ts4900-fpga package
This package is responsible from downloading and deploying bitstream for
TS-4900's FPGA. It implements clocks, UART MUX, and GPIOs. One of these
GPIOs is used to enable the wifi module.

It is loaded by U-Boot during the boot sequence, the default U-boot
script expects to find it in the /boot folder.

Signed-off-by: Mathieu Audat <mathieu.audat@savoirfairelinux.com>
[Thomas:
 - Use "config" instead of "menuconfig" since there's a single option
   and no sub-options.
 - Add BR2_arm dependency, since it's only used on a specific i.MX6
   platform.
 - Implement TS4900_FPGA_EXTRACT_CMDS that copy the file to the package
   build directory $(@D)
 - Change the install command to copy from $(@D) instead of from
   $(DL_DIR)
 - Change the install command to remove the 'mkdir -p
   $(TARGET_DIR)/boot' and instead just use the -D option of the install
   program that creates the destination directory if it doesn't exist.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>