configs/beaglebone_qt5: new defconfig
Add a target configuration for the beaglebone that supports the SGX530
graphics acceleration and builds the qt5 demos with the eglfs backend.
The imagination technologies demos and fbv are also included. The Linux
kernel is based on the omap2plus defconfig, a config fragment adds the
required settings for SGX accelerator support.

Signed-off-by: Lothar Felten <lothar.felten@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>