boards: add FriendlyARM Nanopi NEO
The FriendlyARM Nanopi NEO is a 4x4cm² board with an Allwiner H3 SoC:
  - quad-core Cortex-A7 @1.2GHz
  - 256 or 512MiB of DDR
  - uSDCard as only storage option
  - 3x USB 2.0 host (one socket, two on expansion pin-holes)
  - 1x USB 2.0 OTG (also used as power source)
  - 10/100 etehrnet MAC
  - GPIOs, SPI, I2c...

Support for the Nanopi NEO in U-Boot and Linux is very recent, so much
so that we have to use an -rc tag for U-Boot and a special Linux tree.

As for Linux, I pushed a git tree on Github with a single tag that
matches what is currently queued in the sunxi-next queued for 4.10,
based on 4.9-rc3. All those commits are from Maxime's tree, the
maintainer for most sunxi stuff.

This also means that we can't use the Linux headers from the kernel
being built (which is what we usually do) because those report 4.9,
while Buildroot currently knows only of 4.8 at best. So this is what we
use.

Unfortunately, support for the ethernet MAC and the USB OTG are not yet
upstream, but are being actively worked on.

The Nanopi NEO is very similar to the Orangepi PC, so I was able to
scanvenge most of its configuration. ;-)

Thanks Maxime for your help on IRC! :-)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Maxime Ripard <maxime.ripard@free-electrons.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>