configs/mini2440_defconfig: remove defconfig
This defconfig has not been updated in any significant way since 2011,
when Peter moved it to use a mainline kernel. It's now using an
ancient 3.0 kernel, and a vendor-specific U-Boot version that fails to
build with various "undefined reference" errors:

  undefined reference to `show_boot_progress'
  undefined reference to `image_print_contents'
  undefined reference to `wait_ms'

So it's time to get rid of the defconfig for this platform.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>