kvm-unit-tests: really use KVM_UNIT_TESTS_ARCH and fix value on x86-64
The kvm-unit-tests package defines a KVM_UNIT_TESTS_ARCH variable, but
never used it. Due to this, on PowerPC64, we were passing an incorrect
value, causing a build failure.

In addition, once the KVM_UNIT_TESTS_ARCH is used, it appeared that
its value for x86_64 was incorrect, due a typo: it was x86_84 instead
of x86_64.

Fixes:

  http://autobuild.buildroot.net/results/16836c02f4d1ab5df6b2e21a4fcdbbf4a7542f22/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>